Revert "revert custom rails4.rb changes"
This reverts commit 75fecee2082af76d9b2a02012471381be8a0f77c.

+++ lib/language_pack/rails4.rb
@@ -5,6 +5,17 @@ require "language_pack/rails3"
 class LanguagePack::Rails4 < LanguagePack::Rails3
   ASSETS_CACHE_LIMIT = 52428800 # bytes
 
+  ASSET_PATHS = %w[
+    public/packs
+    ~/.yarn-cache
+    ~/.cache/yarn
+  ]
+
+  ASSET_CACHE_PATHS = %w[
+    node_modules
+    tmp/cache/webpacker
+  ]
+
   # detects if this is a Rails 4.x app
   # @return [Boolean] true if it's a Rails 4.x app
   def self.use?
@@ -24,9 +35,15 @@ class LanguagePack::Rails4 < LanguagePack::Rails3
   def default_process_types
     instrument "rails4.default_process_types" do
       super.merge({
-        "web"     => "bin/rails server -p ${PORT:-5000} -e $RAILS_ENV",
+        "web"     => "bin/rails server -p $PORT -e $RAILS_ENV",
         "console" => "bin/rails console"
       })
+    end
+  end
+
+  def build_bundler(default_bundle_without)
+    instrument "rails4.build_bundler" do
+      super
     end
   end
 
@@ -61,10 +78,17 @@ WARNING
   end
 
   def cleanup
-    super
+    # does not call super because it would return if default_assets_cache was missing
+    # child classes should call super and should not use a return statement
     return if assets_compile_enabled?
-    return unless Dir.exist?(default_assets_cache)
-    FileUtils.remove_dir(default_assets_cache)
+
+    puts "Removing non-essential asset cache directories"
+
+    FileUtils.remove_dir(default_assets_cache) if Dir.exist?(default_assets_cache)
+
+    self.class::ASSET_CACHE_PATHS.each do |path|
+      FileUtils.remove_dir(path) if Dir.exist?(path)
+    end
   end
 
   def run_assets_precompile_rake_task
@@ -76,12 +100,11 @@ WARNING
         end
 
         precompile = rake.task("assets:precompile")
-        return true if precompile.not_defined?
+        return true unless precompile.is_defined?
 
         topic("Preparing app for Rails asset pipeline")
 
-        @cache.load_without_overwrite public_assets_folder
-        @cache.load default_assets_cache
+        load_asset_cache
 
         precompile.invoke(env: rake_env)
 
@@ -89,20 +112,34 @@ WARNING
           log "assets_precompile", :status => "success"
           puts "Asset precompilation completed (#{"%.2f" % precompile.time}s)"
 
-          clean_task = rake.task("assets:clean")
-          if clean_task.task_defined?
-            puts "Cleaning assets"
-            clean_task.invoke(env: rake_env)
+          puts "Cleaning assets"
+          rake.task("assets:clean").invoke(env: rake_env)
 
-            cleanup_assets_cache
-            @cache.store public_assets_folder
-            @cache.store default_assets_cache
-          end
+          cleanup_assets_cache
+          store_asset_cache
         else
           precompile_fail(precompile.output)
         end
       end
     end
+  end
+
+  def load_asset_cache
+    puts "Loading asset cache"
+    @cache.load_without_overwrite public_assets_folder
+    @cache.load default_assets_cache
+
+    paths = (self.class::ASSET_PATHS + self.class::ASSET_CACHE_PATHS)
+    paths.each { |path| @cache.load path }
+  end
+
+  def store_asset_cache
+    puts "Storing asset cache"
+    @cache.store public_assets_folder
+    @cache.store default_assets_cache
+
+    paths = (self.class::ASSET_PATHS + self.class::ASSET_CACHE_PATHS)
+    paths.each { |path| @cache.store path }
   end
 
   def cleanup_assets_cache
